test(settings): split text display coverage
Why:
- Continue migrating package-owned logic tests out of the app-host AndBibleTests bundle.
- Keep Android text-display inheritance and Strong's mode parity covered in the package lanes that own the behavior.

What Changed:
- Added a dedicated BibleCore TextDisplaySettingsTests suite for defaults, inheritance, full resolution, and dirty-override cleanup.
- Added a BibleUI ReaderDisplayConfigTests suite for legacy Strong's mode normalization.
- Removed the migrated app-host tests and moved misplaced text-display cases out of BookmarkModelTests.
- Updated the test architecture migration plan ledger.

Validation:
- git diff --check.
- python3 scripts/check_repo_standards.py docblocks --all-files.
- python3 scripts/check_repo_standards.py source-guards.
- swift package describe --type json.
- xcodebuild test -project AndBible.xcodeproj -scheme BibleCoreTests -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 16,OS=18.6' -derivedDataPath .derivedData/LocalBibleCoreTextDisplayTests CODE_SIGNING_ALLOWED=NO -only-testing:BibleCoreTests/TextDisplaySettingsTests.
- xcodebuild test -project AndBible.xcodeproj -scheme BibleUITests -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 16,OS=18.6' -derivedDataPath .derivedData/LocalBibleUITextDisplayTests CODE_SIGNING_ALLOWED=NO -only-testing:BibleUITests/ReaderDisplayConfigTests.
- GitNexus detect_changes on the explicit worktree staged diff: low risk, no affected flows.
- Adversarial review found no issues.

Impact:
- Moves five app-host tests into app-host-free package test lanes without changing production behavior.
- Reduces misplaced test coverage in the bookmark model suite.
- Risk is low and limited to test organization.
