+++ AndBibleTests/AndBibleTests+AppAndReader.swift
@@ -19,99 +19,6 @@ import struct SwiftUI.Color
 #endif
 
 extension AndBibleTests {
-    func testTextDisplayAppDefaultsUseAndroidHiddenLinksStrongsMode() {
-        XCTAssertEqual(TextDisplaySettings.appDefaults.strongsMode, 0)
-    }
-
-    func testReaderDisplayConfigNormalizesLegacyStrongsModeToAndroidHiddenLinks() {
-        var settings = TextDisplaySettings()
-        settings.strongsMode = 3
-
-        let config = BibleReaderDisplayConfig(settings: settings, defaults: .appDefaults)
-
-        XCTAssertEqual(config.strongsMode, StrongsMode.hiddenLinks.rawValue)
-    }
-
-    func testTextDisplaySettingsInheritanceUsesGlobalBeforeDefaults() {
-        var windowSettings = TextDisplaySettings()
-        windowSettings.fontSize = 18
-
-        var workspaceSettings = TextDisplaySettings()
-        workspaceSettings.fontSize = 16
-        workspaceSettings.fontFamily = "serif"
-
-        var globalSettings = TextDisplaySettings()
-        globalSettings.lineSpacing = 125
-
-        var defaults = TextDisplaySettings()
-        defaults.fontSize = 14
-        defaults.fontFamily = "sans-serif"
-        defaults.lineSpacing = 150
-
-        XCTAssertEqual(
-            TextDisplaySettings.resolved(
-                \.fontSize,
-                window: windowSettings,
-                workspace: workspaceSettings,
-                global: globalSettings,
-                defaults: defaults
-            ),
-            18
-        )
-        XCTAssertEqual(
-            TextDisplaySettings.resolved(
-                \.fontFamily,
-                window: windowSettings,
-                workspace: workspaceSettings,
-                global: globalSettings,
-                defaults: defaults
-            ),
-            "serif"
-        )
-        XCTAssertEqual(
-            TextDisplaySettings.resolved(
-                \.lineSpacing,
-                window: windowSettings,
-                workspace: workspaceSettings,
-                global: globalSettings,
-                defaults: defaults
-            ),
-            125
-        )
-        XCTAssertNil(
-            TextDisplaySettings.resolved(
-                \.topMargin,
-                window: windowSettings,
-                workspace: workspaceSettings,
-                global: globalSettings,
-                defaults: defaults
-            )
-        )
-    }
-
-    func testTextDisplaySettingsFullyResolvedUsesGlobalBeforeDefaults() {
-        let dayBackground = Int(Int32(bitPattern: 0xFFFAF4E8))
-        let nightTextColor = Int(Int32(bitPattern: 0xFFF1E7D0))
-        let workspaceNightTextColor = Int(Int32(bitPattern: 0xFFCCCCCC))
-
-        var globalSettings = TextDisplaySettings()
-        globalSettings.dayBackground = dayBackground
-        globalSettings.nightTextColor = nightTextColor
-
-        var workspaceSettings = TextDisplaySettings()
-        workspaceSettings.nightTextColor = workspaceNightTextColor
-
-        let resolved = TextDisplaySettings.fullyResolved(
-            window: nil,
-            workspace: workspaceSettings,
-            global: globalSettings
-        )
-
-        XCTAssertEqual(resolved.dayBackground, dayBackground)
-        XCTAssertEqual(resolved.nightTextColor, workspaceNightTextColor)
-        XCTAssertEqual(resolved.dayTextColor, TextDisplaySettings.appDefaults.dayTextColor)
-    }
-
     /**
      Protects the extracted SWORD setup boundary for installed-module catalog and active-module
      resolution.
@@ -208,37 +115,6 @@ extension AndBibleTests {
         XCTAssertFalse(manager.isGlobalOptionEnabled(.morphology))
         XCTAssertFalse(manager.isGlobalOptionEnabled(.footnotes))
         XCTAssertFalse(manager.isGlobalOptionEnabled(.crossReferences))
-    }
-
-    func testTextDisplaySettingsChangedFieldsOnlyClearMatchingDirtyOverrides() {
-        var previousGlobal = TextDisplaySettings()
-        previousGlobal.fontSize = 18
-        previousGlobal.lineSpacing = 10
-        previousGlobal.showVerseNumbers = true
-
-        var currentGlobal = previousGlobal
-        currentGlobal.fontSize = 20
-        currentGlobal.showVerseNumbers = false
-
-        let changedFields = TextDisplaySettings.changedFields(
-            from: previousGlobal,
-            to: currentGlobal
-        )
-
-        var childOverrides = TextDisplaySettings()
-        childOverrides.fontSize = 20
-        childOverrides.lineSpacing = 10
-        childOverrides.showVerseNumbers = false
-
-        XCTAssertTrue(
-            childOverrides.clearOverridesMatchingParent(
-                currentGlobal,
-                only: changedFields
-            )
-        )
-        XCTAssertNil(childOverrides.fontSize)
-        XCTAssertNil(childOverrides.showVerseNumbers)
-        XCTAssertEqual(childOverrides.lineSpacing, 10)
     }
 
     func testReaderWindowControlsAvoidanceInsetsStayOffForFullscreenIPad() {
